+++ backend/app/schemacms/projects/tests/test_views.py
@@ -57,9 +57,11 @@ class TestListCreateProjectView:
             == projects_serializers.ProjectSerializer(self.__sort_projects(user1_projects), many=True).data
         )
 
-    def test_create_as_admin(self, api_client, user):
+    def test_create_as_admin(self, api_client, user, mocker):
         user.role = user_constants.UserRole.ADMIN
         api_client.force_authenticate(user)
+
+        create_xlm_file_mock = mocker.patch("schemacms.projects.models.Project.create_xlm_file")
 
         self.example_project["editors"].append(user.id)
         response = api_client.post(self.get_url(), data=self.example_project)
@@ -68,18 +70,23 @@ class TestListCreateProjectView:
 
         assert response.status_code == status.HTTP_201_CREATED
         assert response.data == projects_serializers.ProjectSerializer(instance=project).data
+        create_xlm_file_mock.assert_called_once()
 
     @pytest.mark.parametrize("role", [user_constants.UserRole.EDITOR, user_constants.UserRole.UNDEFINED])
-    def test_create_as_editor(self, api_client, user_factory, role):
+    def test_create_as_editor(self, api_client, user_factory, role, mocker):
         user = user_factory(role=role)
         api_client.force_authenticate(user)
+        create_xlm_file_mock = mocker.patch("schemacms.projects.models.Project.create_xlm_file")
+
         response = api_client.post(self.get_url(), data=self.example_project)
 
         assert response.status_code == status.HTTP_403_FORBIDDEN
+        create_xlm_file_mock.assert_not_called()
 
-    def test_create_without_editors(self, api_client, user):
+    def test_create_without_editors(self, api_client, user, mocker):
         user.role = user_constants.UserRole.ADMIN
         api_client.force_authenticate(user)
+        create_xlm_file_mock = mocker.patch("schemacms.projects.models.Project.create_xlm_file")
 
         self.example_project.pop("editors")
 
@@ -90,6 +97,7 @@ class TestListCreateProjectView:
 
         assert response.status_code == status.HTTP_201_CREATED
         assert response.data == projects_serializers.ProjectSerializer(instance=project).data
+        create_xlm_file_mock.assert_called_once()
 
     def test_num_queries(
         self,
@@ -154,8 +162,9 @@ class TestRetrieveUpdateDeleteProjectView:
         assert "meta" in response.data
         assert response.data["meta"]["data_sources"] == expected
 
-    def test_update_project_by_owner(self, api_client, user, project):
+    def test_update_project_by_owner(self, api_client, user, project, mocker):
         api_client.force_authenticate(user)
+        create_xlm_file_mock = mocker.patch("schemacms.projects.models.Project.create_xlm_file")
 
         new_title = {"title": "new title"}
 
@@ -164,8 +173,9 @@ class TestRetrieveUpdateDeleteProjectView:
         project.refresh_from_db()
         assert response.status_code == status.HTTP_200_OK
         assert response.data == projects_serializers.ProjectSerializer(instance=project).data
+        create_xlm_file_mock.assert_called_once()
 
-    def test_update_project_name_already_occupied(self, api_client, user, project_factory):
+    def test_update_project_name_already_occupied(self, api_client, user, project_factory, mocker):
         project = project_factory()
         other_project = project_factory()
         api_client.force_authenticate(user)
@@ -179,7 +189,7 @@ class TestRetrieveUpdateDeleteProjectView:
             "title": [error.Error(message="This field must be unique.", code="projectTitleUnique").data]
         }
 
-    def test_update_project_by_not_projects_editor(self, api_client, user_factory, project):
+    def test_update_project_by_not_projects_editor(self, api_client, user_factory, project, mocker):
         editor1, editor2 = user_factory.create_batch(2, editor=True)
         project.editors.add(editor2)
         payload = {"title": "new title"}
@@ -190,7 +200,7 @@ class TestRetrieveUpdateDeleteProjectView:
         project.refresh_from_db()
         assert response.status_code == status.HTTP_403_FORBIDDEN
 
-    def test_delete_project_by_owner(self, api_client, user, project):
+    def test_delete_project_by_owner(self, api_client, user, project, mocker):
         api_client.force_authenticate(user)
 
         response = api_client.delete(self.get_url(pk=project.pk))