+++ frontend/schemaCMS/src/modules/page/page.constants.js
@@ -30,6 +30,7 @@ export const PAGE_IS_PUBLIC = 'isPublic';
 export const PAGE_BLOCKS = 'blocks';
 export const PAGE_DELETE_BLOCKS = 'deleteBlocks';
 export const PAGE_TAGS = 'tags';
+export const PAGE_LINK = 'link';
 
 export const BLOCK_NAME = 'name';
 export const BLOCK_TYPE = 'type';
@@ -48,6 +49,7 @@ export const FORM_VALUES = [
   PAGE_DESCRIPTION,
   PAGE_TEMPLATE,
   PAGE_BLOCKS,
+  PAGE_LINK,
   PAGE_IS_PUBLIC,
 ];
 
@@ -59,6 +61,7 @@ export const INITIAL_VALUES = {
   [PAGE_TEMPLATE]: '',
   [PAGE_BLOCKS]: [],
   [PAGE_TAGS]: {},
+  [PAGE_LINK]: '',
   [PAGE_DELETE_BLOCKS]: [],
   [PAGE_IS_PUBLIC]: false,
 };
@@ -113,6 +116,9 @@ export const PAGE_SCHEMA = Yup.object().shape({
   [PAGE_KEYWORDS]: Yup.string()
     .trim()
     .max(1000, 'Page Keywords should have maximum 1000 characters'),
+  [PAGE_LINK]: Yup.string()
+    .trim()
+    .url('The link field needs to be a valid URL'),
   [PAGE_TEMPLATE]: Yup.string()
     .min(1, 'Required')
     .required('Required'),